// resources/frontend/src/app/auth/models/user.ts
export class User {
    id?: number;
    username?: string;
    email?: string;
    password?: string;
    access_token?: string;
    name?: string;
    apellido_paterno?: string;
    apellido_materno?: string;
    avatar?: string;
    is_superuser?: boolean;
    email_verified_at?: Date;
    last_login_at?: Date;
  }